🛡️ Guardian (JULES): Consolidate DiffView components
WHAT:
- Merged src/components/book/elements/diff-view.tsx into src/components/ui/diff-view.tsx
- Updated SideBySideDiff to support WordDiff highlighting and backwards compatibility (before, after props)
- Updated barrel and alias exports to maintain functionality in mdx files
- Removed duplicated LCS computeDiff logic

WHY:
- The book's interactive components implemented an isolated, duplicated VersionDiff and DiffView with its own LCS logic.
- Unifying them under ui/diff-view.tsx reduces maintenance burden.

JULES CONTEXT:
- Checked .Jules/task-log.md: No Autonomous conflicts
- Session logged: .Jules/guardian/2026-05-09/
- Reference: .Jules/JULES.md

IMPACT:
- Files: 2 -> 1
- Import paths: Standardized to @/components/ui/diff-view
- Diff computation consolidated to computeWordDiff

VERIFICATION:
- All unit tests pass
- UI rendering visually backwards compatible in MDX files
