+++ .Jules/task-log.md
@@ -1,0 +1,19 @@
+## 2026-05-09 12:08 - [GUARDIAN] Session started
+- Directory: .Jules/guardian/2026-05-09
+- Phase: PROFILE & SELECT
+- Target: DiffView consolidation
+
+## 2026-05-09 12:40 - [GUARDIAN] Running baseline tests
+- Files: src/components/book/elements/diff-view.tsx src/components/ui/diff-view.tsx
+## 2026-05-09 12:42 - [GUARDIAN] Refactoring
+- Merged VersionDiff into src/components/ui/diff-view.tsx
+- Updated SideBySideDiff to accept before/after/beforeLabel/afterLabel
+- Updated src/components/book/interactive.tsx and src/components/book/elements/index.ts to map DiffView -> SideBySideDiff
+- Deleted src/components/book/elements/diff-view.tsx
+## 2026-05-09 12:44 - [GUARDIAN] Cleanup complete
+- Linting: ✓
+- Formatting: ✓
+## 2026-05-09 12:45 - [GUARDIAN] Verification complete
+- Tests: ✓
+- Types: ✓
+- Build: ✓
